Deployment/requirements.txt
waitress
numpy
streamlit
tensorflow-cpu